Replace boost::remove_if with std::remove_if. Move one seconds_to_date method to the test code since it was not used outside of the tests. Update some comments.
