+++ test/datetime.cc
@@ -211,6 +211,56 @@ void TryIsRestricted(const TimeDomain td, const std::string& date, const bool ex
   }
 }
 
+// Convert seconds to a date string (for test evaluation only). DateTime holds a more general
+// method.
+std::string seconds_to_date(const uint64_t seconds, const boost::local_time::time_zone_ptr& tz) {
+
+  std::string iso_date;
+  if (seconds == 0 || !tz) {
+    return iso_date;
+  }
+
+  try {
+    std::string tz_string;
+    const boost::posix_time::ptime time_epoch(boost::gregorian::date(1970, 1, 1));
+    boost::posix_time::ptime pt = time_epoch + boost::posix_time::seconds(seconds);
+    boost::local_time::local_date_time date_time(pt, tz);
+    pt = date_time.local_time();
+
+    boost::gregorian::date date = pt.date();
+    std::stringstream ss_time;
+    ss_time << pt.time_of_day();
+    std::string time = ss_time.str();
+
+    std::size_t found = time.find_last_of(':'); // remove seconds.
+    if (found != std::string::npos) {
+      time = time.substr(0, found);
+    }
+
+    ss_time.str("");
+    if (date_time.is_dst()) {
+      ss_time << tz->dst_offset() + tz->base_utc_offset();
+    } else {
+      ss_time << tz->base_utc_offset();
+    }
+
+    // positive tz
+    if (ss_time.str().find('+') == std::string::npos &&
+        ss_time.str().find('-') == std::string::npos) {
+      iso_date = to_iso_extended_string(date) + "T" + time + "+" + ss_time.str();
+    } else {
+      iso_date = to_iso_extended_string(date) + "T" + time + ss_time.str();
+    }
+
+    found = iso_date.find_last_of(':'); // remove seconds.
+    if (found != std::string::npos) {
+      iso_date = iso_date.substr(0, found);
+    }
+
+  } catch (std::exception& e) {}
+  return iso_date;
+}
+
 void TryTestTimezoneDiff(const bool is_depart,
                          const uint64_t date_time,
                          const std::string& expected1,
@@ -223,18 +273,18 @@ void TryTestTimezoneDiff(const bool is_depart,
   auto tz1 = DateTime::get_tz_db().from_index(DateTime::get_tz_db().to_index(time_zone1));
   auto tz2 = DateTime::get_tz_db().from_index(DateTime::get_tz_db().to_index(time_zone2));
 
-  std::cout << DateTime::seconds_to_date(dt, tz1) << std::endl;
+  std::cout << seconds_to_date(dt, tz1) << std::endl;
 
   dt += DateTime::timezone_diff(is_depart, dt, tz1, tz2);
-  if (DateTime::seconds_to_date(dt, tz1) != expected1)
+  if (seconds_to_date(dt, tz1) != expected1)
     throw std::runtime_error("Timezone Diff test #1: " + std::to_string(date_time) +
                              " test failed.  Expected: " + expected1 + " but got " +
-                             DateTime::seconds_to_date(dt, tz1));
+                             seconds_to_date(dt, tz1));
 
-  if (DateTime::seconds_to_date(dt, tz2) != expected2)
+  if (seconds_to_date(dt, tz2) != expected2)
     throw std::runtime_error("Timezone Diff test #2: " + std::to_string(date_time) +
                              " test failed.  Expected: " + expected2 + " but got " +
-                             DateTime::seconds_to_date(dt, tz2));
+                             seconds_to_date(dt, tz2));
 }
 
 } // namespace
